+++ ChangeLog.md
@@ -4,6 +4,14 @@
 
 
 
+**版本介绍**
+
+- **x.y.z**     	保证的兼容性，可新增功能（用于版本阶段性开发）、修复或调整版本。（待移除使用 `// Deprecated:  descript text` 标记，或说明）
+- **x.y**            不保证的兼容性，删除旧版本遗弃的方法
+- **x**               重大（颠覆性）的改变，重要里程碑开发
+
+
+
 ## 1.0.x
 
 - [ ] 删除历史版本中标注遗弃的方法
@@ -37,16 +45,25 @@
   - +)  `BitSize` 增加方法 `Bit/Byte` 以完善可自由获取不同的单位
   - optimize) `BitSize`默认输出采用 1024 制式而非原来的1000
   - +) `Unit` 增加方法 `Unit` 以完善可自由获取不同的单位
+  - +) `Bytes` 方法用于字节快速转 `BitSize` 的方法
+  
 - **bin**
+  
   - +) `StructToMap` 和 `StructToMapLStyle` 新增 struct 转 map 的方法
   - 优化) `FormatQue` 使用反射调整其支持任一类型的 array/slice数据
+  
 - **util**
   - +) `InQueAny` 新增方法来支持任一类型的 array/slice数据键值检测
+  
 - **str**
   - +) `IsLatinAlpha` 判断是否为拉丁文字母
   - +) `LowerStyle` 将峰坨式命名规则转为小写风格命名规范
   - +) `CamelCase`  使用 snake case（蛇式）转 CamelCase 实现
-  
+
+- **fs**
+  - +) `GetMemStatsRange` 新增方法用于计算程序内存消耗
+  - +) `MemUsage` 新增结构体用于快速计算系统的内存消耗情况
+
   
 
 ### 1.0.1/2021-06-11